fix active window logic
This makes some changes to the way that the active window is retrieved.

The current logic made some assumptions around the hierarchy expected
for the UIWindow that didn't always hold true.  This instead delays
retrieving the active window until the image cropper is about to be
launched (because the active window may have changed between plugin
initialization and the time where we go and display the image
cropper).

Additionally it iterates over the list of windows to find the first
window that isn't hidden to ensure that we're not trying to present the
image cropper on a hidden window.